\name{roll_collapse}
\alias{roll_collapse}
\title{Rolling collapse}
\usage{
  roll_collapse(x, n)
}
\arguments{
  \item{x}{an R object of form: character vector, character
  matrix}

  \item{n}{an integer; number of elements to 'roll' over.}
}
\description{
  This function implements a rolling collapse with
  C++/Rcpp.
}

