\name{roll_median}
\alias{roll_median}
\title{Rolling median}
\usage{
  roll_median(x, n)
}
\arguments{
  \item{x}{an \R object of form: numeric vector, numeric
  matrix}

  \item{n}{an integer; number of elements to 'roll' over.}
}
\description{
  This function implements a rolling median with C++/Rcpp.
}

