\name{roll_stddev}
\alias{roll_stddev}
\title{Rolling stddev}
\usage{
  roll_stddev(x, n)
}
\arguments{
  \item{x}{an \R object of form: numeric vector, numeric
  matrix}

  \item{n}{an integer; number of elements to 'roll' over.}
}
\description{
  This function implements a rolling stddev with C++/Rcpp.
}

