% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RcppRoll-exports}
\alias{RcppRoll-exports}
\alias{roll_max}
\alias{roll_maxl}
\alias{roll_maxr}
\alias{roll_mean}
\alias{roll_meanl}
\alias{roll_meanr}
\alias{roll_median}
\alias{roll_medianl}
\alias{roll_medianr}
\alias{roll_min}
\alias{roll_minl}
\alias{roll_minr}
\alias{roll_prod}
\alias{roll_prodl}
\alias{roll_prodr}
\alias{roll_sd}
\alias{roll_sdl}
\alias{roll_sdr}
\alias{roll_sum}
\alias{roll_suml}
\alias{roll_sumr}
\alias{roll_var}
\alias{roll_varl}
\alias{roll_varr}
\title{RcppRoll}
\usage{
roll_mean(x, n, weights = NULL, by = 1L, fill = numeric(0),
  partial = FALSE, align = c("center", "left", "right"), normalize = TRUE)

roll_meanr(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "right", normalize = TRUE)

roll_meanl(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "left", normalize = TRUE)

roll_median(x, n, weights = NULL, by = 1L, fill = numeric(0),
  partial = FALSE, align = c("center", "left", "right"), normalize = TRUE)

roll_medianr(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "right", normalize = TRUE)

roll_medianl(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "left", normalize = TRUE)

roll_min(x, n, weights = NULL, by = 1L, fill = numeric(0),
  partial = FALSE, align = c("center", "left", "right"), normalize = TRUE)

roll_minr(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "right", normalize = TRUE)

roll_minl(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "left", normalize = TRUE)

roll_max(x, n, weights = NULL, by = 1L, fill = numeric(0),
  partial = FALSE, align = c("center", "left", "right"), normalize = TRUE)

roll_maxr(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "right", normalize = TRUE)

roll_maxl(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "left", normalize = TRUE)

roll_prod(x, n, weights = NULL, by = 1L, fill = numeric(0),
  partial = FALSE, align = c("center", "left", "right"), normalize = TRUE)

roll_prodr(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "right", normalize = TRUE)

roll_prodl(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "left", normalize = TRUE)

roll_sum(x, n, weights = NULL, by = 1L, fill = numeric(0),
  partial = FALSE, align = c("center", "left", "right"), normalize = TRUE)

roll_sumr(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "right", normalize = TRUE)

roll_suml(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "left", normalize = TRUE)

roll_sd(x, n, weights = NULL, by = 1L, fill = numeric(0),
  partial = FALSE, align = c("center", "left", "right"), normalize = TRUE)

roll_sdr(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "right", normalize = TRUE)

roll_sdl(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "left", normalize = TRUE)

roll_var(x, n, weights = NULL, by = 1L, fill = numeric(0),
  partial = FALSE, align = c("center", "left", "right"), normalize = TRUE)

roll_varr(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "right", normalize = TRUE)

roll_varl(x, n, weights = NULL, by = 1L, fill = NA, partial = FALSE,
  align = "left", normalize = TRUE)
}
\arguments{
\item{x}{A numeric vector or a numeric matrix.}

\item{n}{A window size.}

\item{weights}{A vector of length \code{n}, giving the weights for each
element within a window. If \code{NULL}, we take unit weights of width \code{n}.}

\item{by}{Calculate at every \code{by}-th point rather than every point.}

\item{fill}{Either an empty vector (no fill), or a vector (recycled to)
length 3 giving left, middle and right fills.}

\item{partial}{Partial application? Currently unimplemented.}

\item{align}{Align windows on the \code{"left"}, \code{"middle"} or
\code{"right"}.}

\item{normalize}{Normalize window weights, such that they sum to \code{n}.}
}
\description{
The following functions are provided directly by \code{RcppRoll}.
}

