% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppRoll-package.R
\docType{package}
\name{RcppRoll-package}
\alias{RcppRoll}
\alias{RcppRoll-package}
\title{RcppRoll}
\description{
This package implements a number of 'roll'-ing functions for \R
vectors and matrices.
}
\details{
Currently, the exported functions are:
 \itemize{
 \item{\code{\link{roll_max}}}
 \item{\code{\link{roll_mean}}}
 \item{\code{\link{roll_median}}}
 \item{\code{\link{roll_min}}}
 \item{\code{\link{roll_prod}}}
 \item{\code{\link{roll_sd}}}
 \item{\code{\link{roll_sum}}}
 \item{\code{\link{roll_var}}}
 }
}
\section{Parallelization}{


When the package is compiled with \code{OpenMP} support, the rolling
window computations are parallelized across threads. By default, the
number of threads is chosen by the \code{OpenMP} runtime (typically
controlled through the \code{OMP_NUM_THREADS} environment variable);
it can be set explicitly with \code{options(RcppRoll.threads = <n>)},
and parallelization can be disabled with
\code{options(RcppRoll.threads = 1)}. Small inputs are always computed
serially, and results are identical whatever the number of threads --
including on builds without \code{OpenMP} support at all.

Use \code{\link{roll_threads}()} to check how many threads are in use,
or whether the installed package has \code{OpenMP} support at all. The
package also reports this when attached; suppress that message with
\code{options(RcppRoll.quiet = TRUE)}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kevinushey/RcppRoll}
  \item Report bugs at \url{https://github.com/kevinushey/RcppRoll/issues}
}

}
\author{
\strong{Maintainer}: Kevin Ushey \email{kevinushey@gmail.com}

}
