% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{RcppRoll}
\alias{RcppRoll}
\alias{RcppRoll-package}
\alias{roll_max}
\alias{roll_mean}
\alias{roll_median}
\alias{roll_min}
\alias{roll_prod}
\alias{roll_sd}
\alias{roll_sum}
\alias{roll_var}
\title{RcppRoll}
\usage{
roll_mean(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_median(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_min(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_max(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_prod(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_sum(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_sd(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_var(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_mean(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_median(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_min(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_max(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_prod(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_sum(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_sd(x, n, weights = NULL, pad = NULL, normalize = FALSE)

roll_var(x, n, weights = NULL, pad = NULL, normalize = FALSE)
}
\description{
This package implements a number of 'roll'-ing functions for \R
vectors and matrices.
}
\details{
Currently, the exported functions are:
 \itemize{
  \item{\code{\link{roll_max}}}
 \item{\code{\link{roll_mean}}}
 \item{\code{\link{roll_median}}}
 \item{\code{\link{roll_min}}}
 \item{\code{\link{roll_prod}}}
 \item{\code{\link{roll_sd}}}
 \item{\code{\link{roll_sum}}}
 \item{\code{\link{roll_var}}}
 }
}
\seealso{
\code{\link{rollit}} for 'roll'-ing your own custom functions.
}

