% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{RcppRoll}
\alias{RcppRoll}
\alias{RcppRoll-package}
\title{RcppRoll}
\description{
This package implements a number of 'roll'-ing functions for \R
vectors and matrices.
}
\details{
Currently, the exported functions are:
 \itemize{
 \item{\code{\link{roll_max}}}
 \item{\code{\link{roll_mean}}}
 \item{\code{\link{roll_median}}}
 \item{\code{\link{roll_min}}}
 \item{\code{\link{roll_prod}}}
 \item{\code{\link{roll_sd}}}
 \item{\code{\link{roll_sum}}}
 \item{\code{\link{roll_var}}}
 }
}
\seealso{
\code{\link{rollit}} for 'roll'-ing your own custom functions.
}

