% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rollit_example}
\alias{rollit_example}
\title{'rollit' Output -- Example Function}
\usage{
rollit_example(x, n, by.column, weights, normalize = FALSE)
}
\arguments{
\item{x}{A vector/matrix of numeric type.}

\item{n}{integer; the window / subset size to roll over.}

\item{by.column}{boolean; if \code{TRUE} we loop over columns,
otherwise we loop over rows.}

\item{weights}{a vector of length \code{n}; specify the weighting
to assign to each element in the window.}

\item{normalize}{boolean; if \code{TRUE} we normalize the weights to
sum to 1.}
}
\value{
This function does not return anything; it merely exists
as a skeleton to provide documentation for your own \code{rollit}
generated functions.
}
\description{
This presents the function signature for the output of
\code{\link{rollit}}.
}
\note{
Elements in \code{weights} equal to 0 are skipped, so that e.g.
setting \code{weights = c(1,0,1,0,1)} would skip every 2nd and 4th
element in each length-5 window.
}
\seealso{
\code{\link{rollit}}
}

